[Win10][filesystem][smb] Cleaning up #includes and namespace usings Removed unused functions. Added CustomBuffer.h to share the CustomBuffer implementation between WinLibraryFile.cpp and UwpSmbFile.cpp
