+++ xbmc/platform/win10/filesystem/WinLibraryFile.cpp
@@ -9,6 +9,7 @@
 #include "WinLibraryFile.h"
 #include "WinLibraryDirectory.h"
 #include "platform/win10/AsyncHelpers.h"
+#include "platform/win10/CustomBuffer.h"
 #include "platform/win32/CharsetConverter.h"
 #include "platform/win32/WIN32Util.h"
 #include "utils/log.h"
@@ -27,40 +28,14 @@
 
 using namespace XFILE;
 using namespace KODI::PLATFORM::WINDOWS;
-namespace winrt
-{
-  using namespace Windows::Foundation;
-}
 using namespace winrt::Windows::ApplicationModel;
+using namespace winrt::Windows::Foundation;
 using namespace winrt::Windows::Foundation::Collections;
 using namespace winrt::Windows::Security::Cryptography;
 using namespace winrt::Windows::Storage;
 using namespace winrt::Windows::Storage::AccessCache;
 using namespace winrt::Windows::Storage::Search;
 using namespace winrt::Windows::Storage::Streams;
-
-struct __declspec(uuid("905a0fef-bc53-11df-8c49-001e4fc686da")) IBufferByteAccess : ::IUnknown
-{
-  virtual HRESULT __stdcall Buffer(void** value) = 0;
-};
-
-struct CustomBuffer : winrt::implements<CustomBuffer, IBuffer, IBufferByteAccess>
-{
-  void *m_address;
-  uint32_t m_capacity;
-  uint32_t m_length;
-
-  CustomBuffer(void *address, uint32_t capacity) : m_address(address), m_capacity(capacity), m_length(0) { }
-  uint32_t Capacity() const { return m_capacity; }
-  uint32_t Length() const { return m_length; }
-  void Length(uint32_t length) { m_length = length; }
-
-  HRESULT __stdcall Buffer(void** value) final
-  {
-    *value = m_address;
-    return S_OK;
-  }
-};
 
 CWinLibraryFile::CWinLibraryFile() = default;
 CWinLibraryFile::~CWinLibraryFile(void) = default;
